# PythonRational optimizations