These are the optimizations suggested at https://github.com/sympy/sympy/issues/12602. That issue is still incomplete because I would like to also see these things ported to the core (but likely I won't do that in this PR).