@asmeurer 2017-05-19
Oh, I didn't actually merge this branch with master correctly, which gives an interesting set of benchmarks (all in Python 3.5):

**master**(remember in master igcd_lehmer and math.gcd isn't integrated with PythonRational at all)
```
In [3]: %time count_roots(expr)
CPU times: user 1min 37s, sys: 398 ms, total: 1min 37s
Wall time: 1min 37s
```

**this branch, before merging master** (the time here is coming from the igcd cache)
```
In [2]: %time count_roots(expr)
CPU times: user 15.9 s, sys: 309 ms, total: 16.2 s
Wall time: 16.5 s
```

**d20bce6 (just the commit adding python_gcd to `PythonRational.__init__`) merged with master**
```
In [2]: %time count_roots(expr)
CPU times: user 5.27 s, sys: 22.9 ms, total: 5.29 s
Wall time: 5.31 s
```

**this branch, after merging master**
```
In [2]: %time count_roots(expr)
CPU times: user 2.08 s, sys: 14.5 ms, total: 2.1 s
Wall time: 2.1 s
```

So the igcd cache resulted in a 5.9x speedup, the math.gcd resulted in a 3.1x speedup, and the crosswise optimizations resulted in a 2.5x speedup. 

Now, the bad news, is that in Python 2, we don't see the math.gcd/igcd_lehmer speedup at all. I guess the numbers are too small to use igcd_lehmer. The times go `1min 37s`, `15.4 s`, `15.8 s`, and `5.24 s`, respectively. 

And of course, it only takes 304 ms when gmpy2 is installed.