@jksuom 2017-05-20
This looks good. I did try to read carefully the new code lines.