@asmeurer 2017-06-25
OK, I pushed up some fixes. I had to allow Rational and Integer input in PythonRational to make the removal of `as_int` work (the code in ring_series, mentioned in the [above discussion](https://github.com/sympy/sympy/pull/12656#discussion_r117604184), assumes that this works, and I figured this was easier than trying to add cases to ring_series to not do that).