@smichr 2017-06-28
@jksuom said it looked good up to the last commits and the new additions look fine. I agree that the caller should make the args to igcd ints. The docstring also states that it is for Python ints.